Updating the "port" argument of various functions/classes.
Updating it to be an uint16_t instead of a string.

+++ Include/discordcoreapi/FoundationEntities.hpp
@@ -61,7 +61,7 @@ namespace DiscordCoreInternal {
 	struct DiscordCoreAPI_Dll VoiceSocketProtocolPayloadData {
 		std::string voiceEncryptionMode{};
 		std::string externalIp{};
-		uint64_t voicePort{};
+		uint16_t voicePort{};
 
 		operator DiscordCoreAPI::Jsonifier();
 	};
